Fixed compiler issue with single imports
