+++ src/actions/actionCreatorsTodoList.js
@@ -21,19 +21,4 @@ const setFilter = filter => {
   };
 };
 
-//unsure if this is an actual reducer
-//this is more like a helper function, but the structure is similar to a reducer so I'll leave it here
-const getVisibleTodos = (todos, filter) => {
-  switch (filter) {
-    case "SHOW_ALL":
-      return todos;
-    case "SHOW_ACTIVE":
-      return todos.filter(t => !t.completed);
-    case "SHOW_COMPLETED":
-      return todos.filter(t => t.completed);
-    default:
-      return todos;
-  }
-};
-
-export { addTodo, visibleTodo, setFilter, getVisibleTodos };
+export { addTodo, visibleTodo, setFilter };
