+++ src/components/presentational/Footer.js
@@ -1,0 +1,14 @@
+import React from "react";
+import FilterLink from "../container/FilterLink";
+
+const Footer = () => (
+  <p>
+    Show: <FilterLink filter="SHOW_ALL">All</FilterLink>
+    {", "}
+    <FilterLink filter="SHOW_ACTIVE">Active</FilterLink>
+    {", "}
+    <FilterLink filter="SHOW_COMPLETED">Completed</FilterLink>
+  </p>
+);
+
+export default Footer;
