+++ src/index.js
@@ -1,9 +1,2 @@
-import React from "react";
-import ReactDOM from "react-dom";
-
 import CounterList from "./CounterList";
-
-//import Counter from "./Counter";
-
-const rootElement = document.getElementById("root2");
-ReactDOM.render(<CounterList />, rootElement);
+CounterList();
